[#10] fixed type of noise chunk size

+++ source/serendipity/app.d
@@ -46,7 +46,7 @@ void startEventLoop(SerendipitySettings* settings, SerendipityLogger logger)
     auto writer = ALSADevice("pulse", true, 32, 16_000);
     auto regressor = Regressor(settings.regressor);
     auto synth = new FluidSynth(settings.soundfont);
-    auto noiseChunkSize = cast(size_t)(settings.entropyRate * chunkSize);
+    auto noiseChunkSize = cast(uint)(settings.entropyRate * chunkSize);
 
     while (reader.readable)
     {
